requirements.txt
einops==0.8.1
imageio==2.31.5
joblib==1.4.2
matplotlib==3.8.0
numpy
numpyencoder==0.3.0
optuna==4.2.1
pandas==2.2.3
scikit_learn==1.6.1
scipy
seaborn
setuptools==75.8.0
smplx==0.1.28
timm==1.0.15
torch_dct==0.1.6
tqdm==4.66.1
trimesh==4.6.8
wandb
gdown
chumpy